requirements.txt
daphne
channels
Django
channels-redis
websocket-client